src/deployment/requirements.txt
# Minimal requirements for deployment service only
# Note: torch and torchvision are installed separately in Dockerfile

# Core libraries - MUST be numpy 1.x for PyTorch compatibility
numpy==1.24.3

# Image processing (headless version for Docker)
opencv-python-headless==4.8.1.78
Pillow==9.5.0

# API and web
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==1.10.13
python-multipart==0.0.6

# Machine learning (needed by training module imports)
scikit-learn==1.3.2
mlflow==2.8.1

# Visualization (needed by training module imports)
matplotlib==3.7.5
seaborn==0.12.2

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
psutil>=5.9.0
